clean utils
