+++ .env.tmpl
@@ -1,5 +1,7 @@
 # dot env file for docker-compose
 
-MYSQL_ROOT_PASSWORD=[password]
+MYSQL_ROOT_PASSWORD=[root-password]
 MYSQL_USER=westlab
 MYSQL_PASSWORD=[password]
+
+DSN=westlab:[password]@tcp(mysqld)/glory_test
