+++ golang/db/model.go
@@ -1,4 +1,4 @@
-package utils
+package db
 
 import "time"
